Rebuild as a wrapper around IO::Socket::Async::SSL
This should improve stability, as `IO::Socket::Async::SSL` has seen more
polishing (the basis of Cro) than `IO::Socket::SSL` based on the
OpenSSL` module, which has known segfault and endless loop issues.

This improves maintainance as well, as there is only one OpenSSL stack (and
this wrapper) to maintain instead of two stacks (the synchronous and the
asynchronous one).